requirements.txt
Flask==3.0.0
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
flower==2.0.1
pytest==7.4.3